removes DirectIdentifiableStore.java
