+++ backend/src/main/java/com/bakdata/conquery/io/storage/IdentifiableStore.java
@@ -17,7 +17,7 @@ import lombok.experimental.Accessors;
 @Accessors(fluent = true)
 @Setter
 @Getter
-public abstract class IdentifiableStore<VALUE extends Identifiable<?>> extends KeyIncludingStore<Id<VALUE>, VALUE> {
+public class IdentifiableStore<VALUE extends Identifiable<?>> extends KeyIncludingStore<Id<VALUE>, VALUE> {
 
 	// TODO: 09.01.2020 fk: Consider making these part of a class that is passed on creation instead so they are less loosely bound.
 	@NonNull
@@ -32,12 +32,56 @@ public abstract class IdentifiableStore<VALUE extends Identifiable<?>> extends K
 		super(store);
 	}
 
+
 	@Override
-	protected abstract Id<VALUE> extractKey(VALUE value);
+	protected Id<VALUE> extractKey(VALUE value) {
+		return (Id<VALUE>) value.getId();
+	}
 
-    @Override
-    protected abstract void removed(VALUE value);
+	@Override
+	protected void removed(VALUE value) {
+		try {
+			if (value == null) {
+				return;
+			}
 
-    @Override
-    protected abstract void added(VALUE value);
+			onRemove.accept(value);
+		}
+		catch (Exception e) {
+			throw new RuntimeException("Failed to remove " + value, e);
+		}
+	}
+
+	@Override
+	protected void added(VALUE value) {
+		try {
+			if (value == null) {
+				return;
+			}
+
+			onAdd.accept(value);
+		}
+		catch (Exception e) {
+			throw new RuntimeException("Failed to add " + value, e);
+		}
+	}
+
+	@Override
+	protected void updated(VALUE value) {
+		try {
+			if (value == null) {
+				return;
+			}
+			final VALUE old = store.get((Id<VALUE>) value.getId());
+
+			if (old != null) {
+				onRemove.accept(old);
+			}
+
+			onAdd.accept(value);
+		}
+		catch (Exception e) {
+			throw new RuntimeException("Failed to add " + value, e);
+		}
+	}
 }
